windows: fixed dealloc webviews logic, implemented takeScreenshot, added WebViewEnvironment and WebViewEnvironmentSettings classes

+++ flutter_inappwebview_windows/windows/types/web_resource_error.cpp
@@ -3,7 +3,7 @@
 
 namespace flutter_inappwebview_plugin
 {
-  WebResourceError::WebResourceError(const std::string& description, const int type)
+  WebResourceError::WebResourceError(const std::string& description, const int64_t type)
     : description(description), type(type)
   {}
